update the Decorator API
